Refactor core identity components to prevent circular imports
